% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O-calibration.R
\name{CalcifTemp}
\alias{CalcifTemp}
\title{Calcification temperature from calcite d18O and d18Ow (inverse of the Paleotemperature Equation)}
\usage{
CalcifTemp(d18Oc, d18Ow)
}
\arguments{
\item{d18Oc}{d18O of calcite on PDB scale}

\item{d18Ow}{d18O of seawater SMOW scale}
}
\value{
Temperature in celcius
}
\description{
Calcification temperature from calcite d18O and d18Ow (inverse of the Paleotemperature Equation)
}
\examples{
CalcifTemp(d18OcFromd18OwTemp(d18Ow = 0.6, 1:26), 0.6)
}
