% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateAge.R
\name{CalibrateAge}
\alias{CalibrateAge}
\title{Calibrate 14C age with Bchron}
\usage{
CalibrateAge(
  df,
  age.14C = "age.14C",
  age.14C.se = "age.14C.se",
  curve = "intcal20",
  return.type = "df",
  offset = 0
)
}
\arguments{
\item{df}{dataframe}

\item{age.14C}{Name of column with 14C ages, Default: 'age.14C'}

\item{age.14C.se}{Name of column with 1se 14C age uncertainty, Default: 'age.14C.se'}

\item{curve}{Calibration curve, Default: 'intcal20'}

\item{return.type}{Return the ammended dataframe or additionally the list of PDFs, Default: 'df'}

\item{offset}{Optional offset applied to all 14C ages, Default: 0}
}
\value{
A dataframe or list
}
\description{
Calibrates a set of 14C ages in a dataframe using the BChron package
}
\details{
A wrapper for Bchron::Bchroncalibrate
}
\examples{
# With defaults
dat <- data.frame(age.14C = c(2000, 20000),
                  age.14C.se = c(100, 200))
CalibrateAge(dat)

# Change the calibration
CalibrateAge(dat, curve = "marine13")

# Return the PDFs
cal.lst <- CalibrateAge(dat, curve = "marine13", return = "lst")
with(cal.lst[[2]][[1]][[1]], {plot(ageGrid, densities)})

# Use different column names
dat <- data.frame(radiocarbon.age = c(2000, 20000),
                 se = c(100, 200))
CalibrateAge(dat, age.14C = "radiocarbon.age", age.14C.se = "se")
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Bchron]{BchronCalibrate}}
 \code{\link[stattools]{SummariseEmpiricalPDF}}
}
