% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-functions.R
\name{d18OcFromd18OwTemp}
\alias{d18OcFromd18OwTemp}
\title{Paleotemperature Equation}
\usage{
d18OcFromd18OwTemp(d18Ow, Temp)
}
\arguments{
\item{d18Ow}{oxygen isotope composition of seawater in permille.}

\item{Temp}{temperature in degree celcius.}
}
\value{
the oxygen isotope composition of calcite on the PDB scale.
}
\description{
Calculate the equilibrium oxygen isotope composition (d18O) of calcification
from seawater oxygen isotope composition (d18Ow) and temperature
(Paleotemperature Equation).
}
\details{
This is an implementation of the paleotemperature equation of O’Neil et
al. (1969) and Shackleton (1974). The factor 0.27 was used to convert from
water on the SMOW scale to calcite on the Peedee Belemnite (PDB) scale, as
described in Anand et al. (2003).
}
\examples{
  d18OcFromd18OwTemp(d18Ow = 0.6, 1:26)
}
\references{
Anand, P., Elderfield, H. and Conte, M. H.: Calibration of Mg/Ca thermometry
in planktonic foraminifera from a sediment trap time series,
Paleoceanography, 18(2), 1050, doi:10.1029/2002PA000846, 2003.

O’Neil, J. R., R. N. Clayton, and T. K. Mayeda, Oxygen isotope fractionation
in divalent metal carbonates, 51, 5547– 5558, 1969.

Shackleton, N. J., Attainment of isotopic equilibrium between ocean water and
the benthonic foraminiferal genus Uvigerina: Isotopic changes in the ocean
during the last glacial, Cent. Natl. Rech. Sci. Colloq. Int., 219, 203–209,
1974.
}
\seealso{
\code{\link{d18OwFromSalinity}}, \code{\link{CalcifTemp}}
}
\author{
Andrew Dolman
}
