% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O-calibration.R
\name{d18OwFromSalinity}
\alias{d18OwFromSalinity}
\title{Calculate d18O of seawater from salinity}
\usage{
d18OwFromSalinity(Salinity)
}
\arguments{
\item{Salinity}{Salinity in grams per litre or parts per thousand.}
}
\value{
d18Ow as permille deviation from the SMOW scale
}
\description{
The equation from Duplessy et al. (1991) based on the
  salinity for water samples collected in the upper 250 m of the Atlantic
  Ocean during the GEOSECS expedition. As reported in Anand et al. 2003
}
\examples{
d18OwFromSalinity(30)
}
\references{
Anand, P., Elderfield, H. and Conte, M. H.: Calibration of Mg/Ca thermometry
in planktonic foraminifera from a sediment trap time series,
Paleoceanography, 18(2), 1050, doi:10.1029/2002PA000846, 2003.

Duplessy, J.-C., L. Labeyrie, A. Juillet-Leclerc, F. Maitre, J. Dupart, and
M. Sarnthein, Surface salinity reconstruction of the north Atlantic Ocean
during the last glacial maximum, Oceanol. Acta, 14, 311– 324, 1991.
}
\seealso{
d18OcFromd18OwTemp CalcifTemp
}
