% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{is.equidistant}
\alias{is.equidistant}
\title{Check for equidistant resolution}
\usage{
is.equidistant(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
a logical value: \code{TRUE} if \code{x} has constant resolution,
  \code{FALSE} otherwise.
}
\description{
This function checks whether a numeric vector (e.g. a depth vector or from a
time series) has equidistant increments, i.e. a constant resolution. This
check includes a numerical tolerance that accounts for the machine
respresentation of floating-point numbers, which circumvents the problems
popular methods of checking equidistance have which use, e.g., \code{sd()} or
\code{unique()} on the difference vector of \code{x}.
}
\examples{

is.equidistant(1 : 10)
is.equidistant(c(2.5, 5, 7.5, 10, 12.5))
is.equidistant(x <- seq(0, 12, 0.1)) # compare this to sd(diff(x)) == 0!

is.equidistant(c(1 : 10, 18))

}
\author{
Andrew Dolman, Thomas Münch
}
