% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lag.R
\name{Lag}
\alias{Lag}
\title{Lag a numeric, character, or factor vector}
\source{
<https://CRAN.R-project.org/package=Hmisc>
}
\usage{
Lag(x, shift = 1)
}
\arguments{
\item{x}{a vector}

\item{shift}{integer specifying the number of observations to be shifted to
the right. Negative values imply shifts to the left.}
}
\value{
a vector like \code{x}.
}
\description{
Shifts a vector \code{shift} elements later. Character or factor variables
are padded with \code{""}, numerics with \code{NA}. The shift may be
negative.
}
\details{
Attributes of the original object are carried along to the new lagged one.
}
\examples{

Lag(1:5,2)
Lag(letters[1:4],2)
Lag(factor(letters[1:4]),-2)
# Find which observations are the first for a given subject
id <- c('a','a','b','b','b','c')
id != Lag(id)
!duplicated(id)

}
\seealso{
\code{\link[stats]{lag}}
}
\author{
Frank E Harrel Jr
}
