% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocatePeaks.R
\name{LocatePeaks}
\alias{LocatePeaks}
\title{Find local maxima or minima}
\source{
This function is adapted from the function \code{which.peaks} in
  the R package \code{mSTEM}:
  <https://cran.r-project.org/web/packages/mSTEM/index.html>

  For the purpose of the prxytools package, the function and its function
  arguments have been renamed.
}
\usage{
LocatePeaks(x, minima = FALSE, partial = FALSE)
}
\arguments{
\item{x}{numeric vector for which maxima (minima) shall be analysed.}

\item{minima}{if \code{FALSE} (the default) locate the positions of the
maxima of \code{x}, else locate the positions of the minima.}

\item{partial}{if \code{TRUE}, the first and last data points of \code{x}
are considered to be maxima or minima as well; defaults to \code{FALSE}.}
}
\value{
a numeric vector of the index positions of the maxima (minima).
}
\description{
Find the positions of local maxima or minima in a sequence.
}
\author{
Zhibing He, Dan Cheng
}
