% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-functions.R
\name{CalcifTemp}
\alias{CalcifTemp}
\title{Inverse Paleotemperature Equation}
\usage{
CalcifTemp(d18Oc, d18Ow)
}
\arguments{
\item{d18Oc}{d18O of calcite in permille on the PDB scale.}

\item{d18Ow}{d18O of seawater in permille on the SMOW scale.}
}
\value{
The calcification temperature in degree Celcius.
}
\description{
Calculate the calcification temperature from the oxygen isotope
compositions of the calcite (d18O) and of the seawater (d18Ow) (inverse of
the Paleotemperature Equation).
}
\examples{
  CalcifTemp(d18OcFromd18OwTemp(d18Ow = 0.6, 1:26), 0.6)
}
\seealso{
\code{\link{d18OwFromSalinity}}, \code{\link{d18OcFromd18OwTemp}}
}
\author{
Andrew Dolman
}
