% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-functions.R
\name{CalibrateAge}
\alias{CalibrateAge}
\title{Calibrate 14C age with Bchron}
\usage{
CalibrateAge(
  df,
  age.14C = "age.14C",
  age.14C.se = "age.14C.se",
  curve = "intcal20",
  return.type = "df",
  offset = NULL
)
}
\arguments{
\item{df}{dataframe of two columns giving the 14C ages and their standard error.}

\item{age.14C}{character; name of the column in \code{df} with the 14C ages;
defaults to \code{"age.14C"}.}

\item{age.14C.se}{character; name of the column in \code{df} with 1 standard
error 14C age uncertainty; defaults to \code{"age.14C.se"}.}

\item{curve}{character; name pf the calibration curve to use; defaults to \code{"intcal20"}.}

\item{return.type}{character; signal the amount of returned information:
return only the ammended dataframe (\code{"df"}) or additionally the list
of PDFs (\code{"lst"}); defaults to (\code{"df"}).}

\item{offset}{character; name of an optional additional column in \code{df}
supplying a variable offset applied to all 14C ages; the default
\code{NULL} means to apply no offset.}
}
\value{
A dataframe or list.
}
\description{
Calibrate a set of 14C ages in a dataframe using the BChron package.
}
\details{
This function is a wrapper for \code{\link[Bchron]{BchronCalibrate}}.
}
\examples{
# With defaults
dat <- data.frame(age.14C = c(2000, 20000),
                  age.14C.se = c(100, 200))
CalibrateAge(dat)

# Change the calibration
CalibrateAge(dat, curve = "marine13")

# Return the PDFs
cal.lst <- CalibrateAge(dat, curve = "marine13", return = "lst")
with(cal.lst[[2]][[1]][[1]], {plot(ageGrid, densities)})

# Use different column names
dat <- data.frame(radiocarbon.age = c(2000, 20000),
                 se = c(100, 200))
CalibrateAge(dat, age.14C = "radiocarbon.age", age.14C.se = "se")
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Bchron]{BchronCalibrate}}
 \code{\link[stattools]{SummariseEmpiricalPDF}}
}
\author{
Andrew Dolman
}
