% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O-calibration.R
\name{d18OcFromd18OwTemp}
\alias{d18OcFromd18OwTemp}
\title{Equilibrium d18O of calcification from d18Ow and Temperature (Paleotemperature Equation)}
\usage{
d18OcFromd18OwTemp(d18Ow, Temp)
}
\arguments{
\item{d18Ow}{d18O of seawater}

\item{Temp}{Temperature in celcius}
}
\value{
d18O of calcite on PDB scale
}
\description{
The paleotemperature equation of O’Neil et al. (1969) and
  Shackleton (1974). The factor 0.27 was used to convert from water on the
  SMOW scale to calcite on the Peedee Belemnite (PDB) scale.

  As described in Anand et al. 2003
}
\examples{
d18OcFromd18OwTemp(d18Ow = 0.6, 1:26)
}
\references{
O’Neil, J. R., R. N. Clayton, and T. K. Mayeda, Oxygen isotope fractionation
in divalent metal carbonates, 51, 5547– 5558, 1969.

Shackleton, N. J., Attainment of isotopic equilibrium between ocean water and
the benthonic foraminiferal genus Uvigerina: Isotopic changes in the ocean
during the last glacial, Cent. Natl. Rech. Sci. Colloq. Int., 219, 203–209,
1974.
}
\seealso{
d18OwFromSalinity CalcifTemp
}
