% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpulseResponse.BergerHeath.R
\name{ImpulseResponse.BergerHeath}
\alias{ImpulseResponse.BergerHeath}
\title{Depth solution for Berger and Heath; e.g. page 65 in Officer and Lynch}
\usage{
ImpulseResponse.BergerHeath(z, d, z0 = 0)
}
\arguments{
\item{z}{depth}

\item{d}{mixing depth}

\item{z0}{epth of the pulse without bioturbation}
}
\value{
weights of the impulse response function
}
\description{
Depth solution for Berger and Heath; e.g. page 65 in Officer and Lynch
}
\author{
Thomas Laepple
}
